requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
pydantic-settings
alembic
pytest
httpx
faker
ruff
anthropic
pandas
python-multipart
pwdlib[bcrypt]
pyjwt
